requirements.txt
pytest
setuptools
pandas
webbot
click
git+ssh://git@github.com/znamlab/flexilims.git#egg=flexilims
pyyaml
tifffile
sphinx
sphinx-rtd-theme
sphinx-click
black
